backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
pydantic>=2.6.0
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0

# Database & migrations
SQLAlchemy>=2.0.25
asyncpg>=0.29.0
alembic>=1.13.1
